+++ adb-20211201/src/Models/DescribeSparkAppDiagnosisInfoRequest.php
@@ -1,0 +1,99 @@
+<?php
+
+// This file is auto-generated, don't edit it. Thanks.
+
+namespace AlibabaCloud\SDK\Adb\V20211201\Models;
+
+use AlibabaCloud\Tea\Model;
+
+class DescribeSparkAppDiagnosisInfoRequest extends Model
+{
+    /**
+     * @description This parameter is required.
+     *
+     * @example s202411061017sh0ad564b000****
+     *
+     * @var string
+     */
+    public $appId;
+
+    /**
+     * @description This parameter is required.
+     *
+     * @example amv-uf663j39b0jd2***
+     *
+     * @var string
+     */
+    public $DBClusterId;
+
+    /**
+     * @description This parameter is required.
+     *
+     * @example PYTHON
+     *
+     * @var string
+     */
+    public $language;
+
+    /**
+     * @description This parameter is required.
+     *
+     * @example cn-hangzhou
+     *
+     * @var string
+     */
+    public $regionId;
+    protected $_name = [
+        'appId'       => 'AppId',
+        'DBClusterId' => 'DBClusterId',
+        'language'    => 'Language',
+        'regionId'    => 'RegionId',
+    ];
+
+    public function validate()
+    {
+    }
+
+    public function toMap()
+    {
+        $res = [];
+        if (null !== $this->appId) {
+            $res['AppId'] = $this->appId;
+        }
+        if (null !== $this->DBClusterId) {
+            $res['DBClusterId'] = $this->DBClusterId;
+        }
+        if (null !== $this->language) {
+            $res['Language'] = $this->language;
+        }
+        if (null !== $this->regionId) {
+            $res['RegionId'] = $this->regionId;
+        }
+
+        return $res;
+    }
+
+    /**
+     * @param array $map
+     *
+     * @return DescribeSparkAppDiagnosisInfoRequest
+     */
+    public static function fromMap($map = [])
+    {
+        $model = new self();
+        if (isset($map['AppId'])) {
+            $model->appId = $map['AppId'];
+        }
+        if (isset($map['DBClusterId'])) {
+            $model->DBClusterId = $map['DBClusterId'];
+        }
+        if (isset($map['Language'])) {
+            $model->language = $map['Language'];
+        }
+        if (isset($map['RegionId'])) {
+            $model->regionId = $map['RegionId'];
+        }
+
+        return $model;
+    }
+}