+++ adb-20211201/src/Models/DescribeSparkAppTypeRequest.php
@@ -1,0 +1,83 @@
+<?php
+
+// This file is auto-generated, don't edit it. Thanks.
+
+namespace AlibabaCloud\SDK\Adb\V20211201\Models;
+
+use AlibabaCloud\Tea\Model;
+
+class DescribeSparkAppTypeRequest extends Model
+{
+    /**
+     * @description This parameter is required.
+     *
+     * @example s202407161205sza4c07c1000****
+     *
+     * @var string
+     */
+    public $appId;
+
+    /**
+     * @description This parameter is required.
+     *
+     * @example am-wz9w49b12933****
+     *
+     * @var string
+     */
+    public $DBClusterId;
+
+    /**
+     * @description This parameter is required.
+     *
+     * @example cn-shenzhen
+     *
+     * @var string
+     */
+    public $regionId;
+    protected $_name = [
+        'appId'       => 'AppId',
+        'DBClusterId' => 'DBClusterId',
+        'regionId'    => 'RegionId',
+    ];
+
+    public function validate()
+    {
+    }
+
+    public function toMap()
+    {
+        $res = [];
+        if (null !== $this->appId) {
+            $res['AppId'] = $this->appId;
+        }
+        if (null !== $this->DBClusterId) {
+            $res['DBClusterId'] = $this->DBClusterId;
+        }
+        if (null !== $this->regionId) {
+            $res['RegionId'] = $this->regionId;
+        }
+
+        return $res;
+    }
+
+    /**
+     * @param array $map
+     *
+     * @return DescribeSparkAppTypeRequest
+     */
+    public static function fromMap($map = [])
+    {
+        $model = new self();
+        if (isset($map['AppId'])) {
+            $model->appId = $map['AppId'];
+        }
+        if (isset($map['DBClusterId'])) {
+            $model->DBClusterId = $map['DBClusterId'];
+        }
+        if (isset($map['RegionId'])) {
+            $model->regionId = $map['RegionId'];
+        }
+
+        return $model;
+    }
+}