+++ adb-20211201/src/Models/DescribeSparkAppTypeResponseBody.php
@@ -1,0 +1,81 @@
+<?php
+
+// This file is auto-generated, don't edit it. Thanks.
+
+namespace AlibabaCloud\SDK\Adb\V20211201\Models;
+
+use AlibabaCloud\Tea\Model;
+
+class DescribeSparkAppTypeResponseBody extends Model
+{
+    /**
+     * @example {
+     * "PolicyType": "AccountLevelIdentityBasedPolicy",
+     * "AuthPrincipalOwnerId": "1*****************7",
+     * "AuthPrincipalType": "SubUser",
+     * "AuthPrincipalDisplayName": "2***************9",
+     * }
+     * @var string
+     */
+    public $accessDeniedDetail;
+
+    /**
+     * @example 596AF63B-8798-501E-BA06-CD2184D48A35
+     *
+     * @var string
+     */
+    public $requestId;
+
+    /**
+     * @example Batch
+     *
+     * @var string
+     */
+    public $type;
+    protected $_name = [
+        'accessDeniedDetail' => 'AccessDeniedDetail',
+        'requestId'          => 'RequestId',
+        'type'               => 'Type',
+    ];
+
+    public function validate()
+    {
+    }
+
+    public function toMap()
+    {
+        $res = [];
+        if (null !== $this->accessDeniedDetail) {
+            $res['AccessDeniedDetail'] = $this->accessDeniedDetail;
+        }
+        if (null !== $this->requestId) {
+            $res['RequestId'] = $this->requestId;
+        }
+        if (null !== $this->type) {
+            $res['Type'] = $this->type;
+        }
+
+        return $res;
+    }
+
+    /**
+     * @param array $map
+     *
+     * @return DescribeSparkAppTypeResponseBody
+     */
+    public static function fromMap($map = [])
+    {
+        $model = new self();
+        if (isset($map['AccessDeniedDetail'])) {
+            $model->accessDeniedDetail = $map['AccessDeniedDetail'];
+        }
+        if (isset($map['RequestId'])) {
+            $model->requestId = $map['RequestId'];
+        }
+        if (isset($map['Type'])) {
+            $model->type = $map['Type'];
+        }
+
+        return $model;
+    }
+}